.cascade/templates/providers/provider_template.py
"""
{provider_name} provider implementation.

This provider {provider_description}
"""
from dataclasses import dataclass
from typing import Optional, Dict, Any

from src.adapters.llm import BaseProvider
from src.types import (
    CodeContext,
    AnalysisResult,
    RefactoringResult,
    ArchitectureResult
)


@dataclass
class {provider_name}Config:
    """Configuration for {provider_name}."""
    api_key: Optional[str] = None
    model: str = "default"
    max_tokens: int = 1000
    temperature: float = 0.7
    additional_params: Dict[str, Any] = None

    def __post_init__(self):
        self.additional_params = self.additional_params or {}


class {provider_name}Provider(BaseProvider):
    """Implementation of {provider_name} provider."""
    
    def __init__(self, config: Optional[{provider_name}Config] = None):
        """Initialize provider with configuration."""
        self.config = config or {provider_name}Config()
        self._client = None
        self._initialized = False
    
    async def initialize(self) -> None:
        """Initialize provider resources."""
        if self._initialized:
            return
        
        try:
            # Initialize provider-specific client/resources here
            self._client = await self._setup_client()
            self._initialized = True
        except Exception as e:
            raise RuntimeError(f"Failed to initialize {self.__class__.__name__}: {str(e)}")
    
    async def close(self) -> None:
        """Clean up provider resources."""
        if self._client:
            # Clean up provider-specific resources here
            await self._client.close()
            self._client = None
        self._initialized = False
    
    async def analyze_code(self, context: CodeContext) -> AnalysisResult:
        """Analyze code with provider."""
        if not self._initialized:
            raise RuntimeError("Provider not initialized")
        
        try:
            # Implement code analysis logic here
            analysis = await self._client.analyze(context.code)
            return AnalysisResult(success=True, analysis=analysis)
        except Exception as e:
            return AnalysisResult(
                success=False,
                error=f"Analysis failed: {str(e)}"
            )
    
    async def suggest_refactoring(self, context: CodeContext) -> RefactoringResult:
        """Suggest code refactoring."""
        if not self._initialized:
            raise RuntimeError("Provider not initialized")
        
        try:
            # Implement refactoring suggestion logic here
            suggestions = await self._client.suggest_refactoring(context.code)
            return RefactoringResult(success=True, suggestions=suggestions)
        except Exception as e:
            return RefactoringResult(
                success=False,
                error=f"Refactoring suggestion failed: {str(e)}"
            )
    
    async def explain_architecture(self, context: CodeContext) -> ArchitectureResult:
        """Explain code architecture."""
        if not self._initialized:
            raise RuntimeError("Provider not initialized")
        
        try:
            # Implement architecture explanation logic here
            explanation = await self._client.explain_architecture(context.code)
            return ArchitectureResult(success=True, explanation=explanation)
        except Exception as e:
            return ArchitectureResult(
                success=False,
                error=f"Architecture explanation failed: {str(e)}"
            )
    
    async def _setup_client(self):
        """Set up provider-specific client."""
        # Implement client setup logic here
        pass
